Reworks so that the Saleforce Product contains Replicated details (#1)
TL;DR
=====

Moves Replicated license details from Opportunity/Contract to Product

Details
=======

After getting this automation working using fields on the Opportunity and
Contract objects, I realized that wasn't quite right. Reps already put the
details of what they're selling onto the Opportunity with the products they
attach, and the Contract isn't really designed to hold that kind of data. So
I set about making it more natural.

The first this change does is to move the details about the entitlements onto
the Product. Then it addresses some confusion I had about how Orders and
Contracts work together, tying the order more tightly to the Replicated
License by saving the license ID on the Order. It keeps the automation for
creating the Order and Contract, but adjusts it so that the Contract gets
created first as expected.

With those two changes, there's now a step to the Contract being approved as
what triggers issuing the license. Now the Contract being activated updates
the Order and that triggers issuing the license.

Also adds a README to the project to help explain what's going on.